requirements.txt
anthropic>=0.42.0
openai>=1.50.0
python-telegram-bot>=21.0
streamlit>=1.40.0
python-dotenv>=1.0.0
Pillow>=10.0.0
python-docx>=1.1.0
